modificaciones en clases
