+++ usuario.json
@@ -1,5 +1,32 @@
-[
-    {"id": 1, "nombre_apellido": "Diego Rodriguez", "dni": "25.356.893", "email": "dr@gmail.com", "telefono": "2494009876"},
-    {"id": 2, "nombre_apellido": "Gisele Keimel", "dni": "32.156.489", "email": "gk@gmail.com", "telefono": "2494005933"},
-    {"id": 3, "nombre_apellido": "Lucas Salvatori", "dni": "27.887.900", "email": "ls@gmail.com", "telefono": "2494673455"}
-]
+{
+    "usuarios": [
+        {
+            "id": 1,
+            "nombre_apellido": "Diego Rodriguez",
+            "dni": "25.356.893",
+            "email": "dr@gmail.com",
+            "telefono": "2494009876"
+        },
+        {
+            "id": 2,
+            "nombre_apellido": "Gisele Keimel",
+            "dni": "32.156.489",
+            "email": "gk@gmail.com",
+            "telefono": "2494005933"
+        },
+        {
+            "id": 3,
+            "nombre_apellido": "Lucas Salvatori",
+            "dni": "27.887.900",
+            "email": "ls@gmail.com",
+            "telefono": "2494673455"
+        },
+        {
+            "id": 4,
+            "nombre_apellido": "laura alvarez",
+            "dni": "435353",
+            "email": "2rfdv",
+            "telefono": "3453"
+        }
+    ]
+}
